general: further warning fixes, use -Wno-unused-variable if NDEBUG
Disabling unused variable checking is needed in NDEBUG builds because of the heavy use of temporary variables in asserts.

+++ dtool/src/dtoolutil/executionEnvironment.cxx
@@ -755,10 +755,10 @@ read_args() {
     if (_binary_name.empty()) {
       _binary_name = buffer;
     }
-    int idx = strlen(buffer) + 1;
+    size_t idx = strlen(buffer) + 1;
     while (idx < bufsize) {
       _args.push_back((char*)(buffer + idx));
-      int newidx = strlen(buffer + idx);
+      size_t newidx = strlen(buffer + idx);
       idx += newidx + 1;
     }
   }
